Base abstract class

All bidsr classes should inherit this abstract class, which provides consistent S3 generics. This is an abstract class, which should not be called directly; use new_bids_class to create new class definitions.

Usage

bids_class_base()