Generate 'BIDS' entity file-name class

Rule-based 'BIDS' entity file names, see class generator at new_bids_entity_file_class; for entity tables; see https://bids-specification.readthedocs.io/en/stable/appendices/entity-table.html

Usage

parse_bids_file_entity_registry(
  definition_table,
  entity_table,
  data_type_table = NULL,
  suffix_table = NULL
)

register_bids_file_entiries(
  definition_table,
  entity_table,
  data_type_table = NULL,
  suffix_table = NULL,
  overwrite = NA
)

bids_entity_file_registry_get(identifier, ifnotfound = NULL, overwrite = NA)

bids_entity_file_registry_list(data_types = NA)

Arguments

definition_table

data table or a path to a 'CSV' file where 'BIDS' entities are defined for each combination of data type and suffix. The columns must start with 'DataType' and 'Suffix', followed by data entities; See 'Examples'.

entity_table

table of two columns: 'Name' (entity keys) and 'Type') entity data type; See 'Examples'.

data_type_table, suffix_table

used to generate descriptions; optional

overwrite

whether to overwrite existing registry if two definitions with the same identifier coincide; default is NA, assuming two definitions are identical, and will not overwrite existing ones; set to TRUE to always force registering the new definitions; set to FALSE and the script will error out if existing definitions have been already registered.

identifier

file registry identifier, automatically derived from data type and suffix; format is 'datatype/suffix'

ifnotfound

default to return is definition is not found

data_types

constraint returns to be under certain data types; default is NA (return the entire registry)

Value

For parse_bids_file_entity_registry, returns a list of entity identifiers and definitions; register_bids_file_entiries registers these definitions to a registry, by identifiers, and returns the number of total definitions registered.

Examples

# ---- Example tables --------------------------------------------

definition_table <- system.file(
  "definitions", "MagneticResonanceImaging.csv",
  package = "bidsr")

print(head(read.csv(definition_table)))
#>   DataType    Suffix      sub      ses     task      acq       ce        rec
#> 1     anat       T1w required optional optional optional optional prohibited
#> 2     anat       T2w required optional optional optional optional prohibited
#> 3     anat       PDw required optional optional optional optional prohibited
#> 4     anat     starw required optional optional optional optional prohibited
#> 5     anat     FLAIR required optional optional optional optional prohibited
#> 6     anat inplaneT1 required optional optional optional optional prohibited
#>        dir        run      mod       echo       flip        inv         mt
#> 1 optional prohibited optional prohibited prohibited prohibited prohibited
#> 2 optional prohibited optional prohibited prohibited prohibited prohibited
#> 3 optional prohibited optional prohibited prohibited prohibited prohibited
#> 4 optional prohibited optional prohibited prohibited prohibited prohibited
#> 5 optional prohibited optional prohibited prohibited prohibited prohibited
#> 6 optional prohibited optional prohibited prohibited prohibited prohibited
#>       part  recording    chunk
#> 1 optional prohibited optional
#> 2 optional prohibited optional
#> 3 optional prohibited optional
#> 4 optional prohibited optional
#> 5 optional prohibited optional
#> 6 optional prohibited optional

entity_table <- system.file(
  "definitions", "MagneticResonanceImagingEntity.csv",
  package = "bidsr")

print(head(read.csv(entity_table)))
#>   Name  Type
#> 1  sub label
#> 2  ses label
#> 3 task label
#> 4  acq label
#> 5   ce label
#> 6  rec label

data_type_table <- system.file(
  "definitions", "glossary", "DataType.csv", package = "bidsr")

print(head(read.csv(data_type_table)))
#>   Name                              FullName
#> 1 anat Anatomical Magnetic Resonance Imaging
#> 2  eeg                Electroencephalography
#> 3 ieeg   Intracranial electroencephalography
#>                                                                                   Description
#> 1 Magnetic resonance imaging sequences designed to characterize static, anatomical features. 
#> 2                                                                      Electroencephalography
#> 3              Intracranial electroencephalography (iEEG) or electrocorticography (ECoG) data

suffix_table <- system.file(
  "definitions", "glossary", "Suffix.csv", package = "bidsr")

print(head(read.csv(suffix_table)))
#>        Name       Description
#> 1       T1w T1-weighted image
#> 2       T2w T2-weighted image
#> 3       PDw                  
#> 4     starw                  
#> 5     FLAIR                  
#> 6 inplaneT1                  

# ---- Parse ------------------------------------------------
parsed <- parse_bids_file_entity_registry(
  definition_table = definition_table,
  entity_table = entity_table,
  data_type_table = data_type_table,
  suffix_table = suffix_table
)

parsed[[1]]
#> $identifier
#> [1] "anat/t1w"
#> 
#> $definition
#> <bidsr::bids_entity_file_anat_t1w> class
#> @ parent     : <bidsr::bids_entity_file>
#> @ constructor: function(parent_directory, sub, ses, task, acq, ce, dir, mod, part, chunk, ..., .list, .suffix, .extension) {...}
#> @ validator  : <NULL>
#> @ properties :
#>  $ parent_directory: <character>
#>  $ data_type       : <character>
#>  $ suffix          : <character>
#>  $ entities        : <list>     
#>  $ extension       : <character>
#>  $ description     : <character>
#>  $ identifier      : <character>
#> 

# ---- Register to registry for querying --------------------
register_bids_file_entiries(
  definition_table = definition_table,
  entity_table = entity_table,
  data_type_table = data_type_table,
  suffix_table = suffix_table
)

# example using definition
t1_definition <- bids_entity_file_registry_get("anat/t1w")

filename <- t1_definition(
  parent_directory = "anat",
  sub = "001",
  ses = "001",
  task = "001",
  acq = "ecog",
  .extension = "nii.gz"
)

# use formatting
format(filename)
#> [1] "anat/sub-001_ses-001_task-001_acq-ecog_T1w.nii.gz"

file.path("BIDS/directory/to", filename)
#> [1] "BIDS/directory/to/anat/sub-001_ses-001_task-001_acq-ecog_T1w.nii.gz"

# with descriptions
filename$print(details = TRUE)
#> [anat/t1w] anat/sub-001_ses-001_task-001_acq-ecog_T1w.nii.gz
#> 	▶ T1-weighted image, Anatomical Magnetic Resonance Imaging

# get subject value
filename$entities$sub$value
#> [1] "001"