'BIDS' subject class

'BIDS' subject class

Usage

bids_subject(project, subject_code, ..., strict = "raw")

Arguments

project

'BIDS' project instance, see bids_project, or a path to the 'BIDS' project

subject_code

character, subject code with or without the leading 'sub-'. The subject code, after trimming the leading entity key, should not contain any additional dash ('-')

...

passed to the constructor of bids_project, when project is a character string

strict

whether to check if the subject folders exist, can be logical or characters; when strict is character strings, choices can be 'raw' (checking raw-data directory) and/or 'source' ( for source-data directory); strict=TRUE is equivalent to checking both; default is 'raw'. There is no checks on derivatives.

Value

A 'BIDS' subject instance.

Examples

example_path <- system.file(
  "examples", "ieeg_epilepsy_ecog", package = "bidsr")

project <- bids_project(
  path = example_path,
  raw_data_relpath = ".",
  derivative_data_relpath = "derivatives"
)

subject <- bids_subject(project = project, subject_code = "ecog01",
                        strict = FALSE)

storage_root <- subject$get_path("raw")

if( dir.exists(storage_root) ) {
  subject$query_modality("ieeg")
}